prevent admin self delete

+++ REST API/src/controllers/Controller.js
@@ -57,6 +57,10 @@ class Controller {
         if (req.user.role === 'admin') {
             try {
                 const id = req.params.id;
+                // Prevent deleting own admin account
+                if(req.user._id === id) return res.status(401).json({ message: "You can\'t delete your user account" })
+
+                // delete entity
                 await this.service.deleteById(id);
                 res.status(204).json({});
             } catch (error) {
